% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSIndex3-package.R
\docType{package}
\name{SSIndex3-package}
\alias{SSIndex3}
\alias{SSIndex3-package}
\title{SSIndex3: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chaoqiw0324/SSIndex3}
  \item Report bugs at \url{https://github.com/chaoqiw0324/SSIndex3/issues}
}

}
\author{
\strong{Maintainer}: Chaoqi Wu \email{cw3370@cumc.columbia.edu} (\href{https://orcid.org/{{orcid}}}{ORCID}) [copyright holder]

}
\keyword{internal}
